requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
decorator==4.4.2
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.2
idna==3.3
imageio==2.11.1
imageio-ffmpeg==0.4.5
moviepy==1.0.3 
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
Pillow==8.4.0
praw==7.5.0
prawcore==2.3.0
proglog==0.1.9
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tqdm==4.62.3
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==1.2.1
